feat: Enhance USB logging functionality and introduce HOLD mode
- Added non-blocking USB logging capabilities to the logger, allowing for asynchronous writes and improved state management.
- Introduced a new HOLD mode in the mode manager to actively maintain the current encoder position.
- Updated navigation controller to check for completion status and log diagnostics to USB.
- Refactored camera vision update functions into a new obstacle_vision.cpp file for better organization.
- Improved stall detection logic to handle DC_HOLDING state separately.
- Enhanced serial handler to switch to HOLD mode via serial command.
- Cleaned up and optimized various functions across multiple files for better performance and readability.

+++ src/position_estimator.cpp
@@ -50,7 +50,7 @@ static float total_distance_traveled = 0;
 static bool pos_initialized = false;
 
 // Slip detection
-static float slip_log_timer = 0;        // Timer for throttle
+static unsigned long slip_log_timer = 0; // micros() timestamp for throttling
 
 // ==========================================
 // INTERNAL HELPERS
